os::syscall: Add memory checks to sys_write (#2)
The existing implementation of sys_write allows a user app to read arbitrary memory 
addresses, which is definitely problematic. This commit adds memory checks to sys_write, 
limiting user apps to reading their own memory ranges and stacks. It is important to note 
that this implementation is specific to the current batch OS.