os::batch: Add a start method
The existing implementation has no limitations on the total size of
apps. Therefore, it is possible for the kernel image to extrude into the
app-reserved addresses. This commit introduces the start method, which
performs a check on the kernel end address before running the first app.
It shuts down the machine if this adverse condition happens.