os::task::runner: Run next task with switch
This commit contains two notable changes. First, run_next_app no longer
return the never type. In fact, except for the first run of an app,
where we crafted a return address, it will now return from the switch
and continue execution there. Second, we no longer need to pass a
pointer to TrapContext to the __restore function, which was previously a
workaround to set the kernel stack pointer during the first app run.
With __switch, this is no longer necessary.