os::syscall: Refactor sys_task_info
The existing implementation prints simple information about the current
task. This commit extends the function so that it now writes the
detailed information of the given task ID into the provided pointer.
Please note that these changes are incompatible with the current
user-side sys_task_info, which will be fixed later.