os::mm::sv39: Fix possible invalid PTE in RootPgt::map_create
The existing implementation may create a new PTE pointing to a physical
page that has been reclaimed if an error occurs during the traversal of
the page table. To fix this issue, we can either roll back the newly
created PTE or push the newly acquired page to the root page table. This
commit implements the latter option because it is assumed that if the
client recovers from the error, they may need to redo the same mapping,
or the process might be killed, and the root table dropped. The latter
option avoids recreating the same PTE.