os::trap: Fix __save_k_ctx in trap.S
The existing implementation is problematic: First, it does
not save the original tp to the saved trap context. Instead,
it puts a zero in the trap context and leaves the original tp
in sscratch. This could be incorrect if the sscratch is
modified afterward, or if another trap occurs. In these cases,
an incorrect tp will be restored. Secondly, it reserves an
incorrect size for trap context (i.e., 34x8), which should be
35x8. This commit fixes these issues.