os::mm::vm: Refactor VMSpace::map_user_elf
In the normal workflow of page mapping, we first find the
VMArea that contains the target page, and then map the page
according to the properties of that VMArea. This fits the
common case of mapping a page at a time. However, when creating
the ELF area, we need to map a bunch of pages with the same
VMArea, and it would be inefficient to search for the VMAre
a for each single page. Therefore, we introduce the area_id.
However, this would cause friction for the mmap we are going
to introduce. This commit refactors the map_user_elf, treating
it as a special case and gradually removes the use of area_id.