os::task: Add update_tcb
It is very strange to have do_page_fault in the task module.
I tried to solve this issue before but failed to come up with
a good idea. This commit introduces a new function update_tcb,
which adds a useful layer of abstraction and can be used to
refactor do_page_fault.