os::task: Refactor get_current_task_id
The existing implementation returns an Option to indicate whether
there is a task running or not. However, most users simply call
expect on the Option, providing slightly different error messages.
This commit changes the return type to usize and panics if no task
is running. try_get_current_task_id, which provides the original
functionality, is added.